move getProcessNamespaceInode
