+++ pkg/util/cgroups/namespace_linux.go
@@ -8,6 +8,12 @@
 
 package cgroups
 
+import (
+	"os"
+	"path/filepath"
+	"syscall"
+)
+
 // From https://github.com/torvalds/linux/blob/5859a2b1991101d6b978f3feb5325dad39421f29/include/linux/proc_ns.h#L41-L49
 // Currently, host namespace inode number are hardcoded, which can be used to detect
 // if we're running in host namespace or not (does not work when running in DinD)
@@ -20,4 +26,21 @@ const (
 func IsProcessHostCgroupNamespace(procPath string, namespaceID uint64) *bool {
 	b := namespaceID == hostCgroupNamespaceInode
 	return &b
+}
+
+// getProcessNamespaceInode performs a stat() call on /proc/<pid>/ns/<namespace>
+//
+// This has been copied from pkg/util/system/namespace_linux.go's GetProcessNamespaceInode. This
+// should ideally live only in pkg/util/system, but currently must be duplicated since the
+// pkg/util/cgroups module needs to be isolated from (not require)
+// github.com/DataDog/datadog-agent, which is the module where pkg/util/system lives.
+func getProcessNamespaceInode(procPath string, pid string, namespace string) (uint64, error) {
+	nsPath := filepath.Join(procPath, pid, "ns", namespace)
+	fi, err := os.Stat(nsPath)
+	if err != nil {
+		return 0, err
+	}
+
+	// We are on linux, casting in safe
+	return fi.Sys().(*syscall.Stat_t).Ino, nil
 }
